renamed typeConfig file
