+++ frontend/src/pages/node/AssociationsTable.vue
@@ -224,7 +224,7 @@ import TableControls from "@/components/TheTableContols.vue";
 import { useQuery } from "@/composables/use-query";
 import { getBreadcrumbs } from "@/pages/node/AssociationsSummary.vue";
 import SectionAssociationDetails from "@/pages/node/SectionAssociationDetails.vue";
-import { fieldFor, TYPE_CONFIG } from "@/util/type-config";
+import { fieldFor, TYPE_CONFIG } from "@/util/typeConfig";
 
 type Props = {
   /** current node */
