+++ frontend/src/util/typeConfig.ts
